fix: Acomodar los archivos
