+++ src/investments/hooks/usePlazoFijo/usePlazoFijoView.ts
@@ -1,0 +1,56 @@
+import { useEffect } from "react";
+import usePaginationParams from "../../../shared/hooks/usePaginateParams";
+import { useCurrentStudent } from "../../../student/hooks/useCurrentStudent";
+import { usePlazoFijoStudent } from "../usePlazoFijoAPI";
+import type { PaginationInfo } from "../../../shared/types/PaginacionType";
+import type { RegisterPlazoFijo } from "../../types/plazoFijo.type";
+
+export const usePlazoFijoView = () => {
+  const {
+    paginationParams,
+    handlePageChange,
+    handlePageSizeChange,
+    // handleFilter,
+  } = usePaginationParams();
+
+  const { plazoFijos, getPaginatedPlazoFijo, registerPlazoFijo } =
+    usePlazoFijoStudent();
+
+  const { wallet } = useCurrentStudent();
+  const userBalance = wallet?.balance;
+
+  useEffect(() => {
+    getPaginatedPlazoFijo(paginationParams);
+  }, [paginationParams]);
+
+  const paginationInfo: PaginationInfo | null = plazoFijos
+    ? {
+        currentPage: plazoFijos.currentPage,
+        totalPages: plazoFijos.totalPages,
+        pageSize: plazoFijos.pageSize,
+        totalItems: plazoFijos.results.length,
+        onPageChange: handlePageChange,
+        onPageSizeChange: handlePageSizeChange,
+      }
+    : null;
+
+  const handleCreatePlazoFijo = async (data: RegisterPlazoFijo) => {
+    registerPlazoFijo(data);
+  };
+
+  return {
+    //Paginacion
+    paginationParams,
+    paginationInfo,
+    handlePageChange,
+    handlePageSizeChange,
+    //Valores del context API
+    plazoFijos,
+    getPaginatedPlazoFijo,
+    registerPlazoFijo,
+
+    //Valores internos
+    userBalance,
+    handleCreatePlazoFijo,
+  };
+};
